+++ src/_modules/messages/index.ts
@@ -1,5 +1,5 @@
 export { MessageService } from './message.service';
 export { MessageController } from './message.controller';
 export { MessageRouter } from './message.router';
-export { ViewMessageDTO } from './message.model';
+export { MessageResponseDTO } from './message.model';
 export type { ICreateMessage } from './message.model';