+++ src/_utils/response_helper.ts
@@ -1,3 +1,14 @@
+import { t, TSchema } from "elysia";
+
 export const swaggerDetails = (title: string, description?: string) => {
-    return {summary: title.toWellFormed(), description: description }
+    return { summary: title.toWellFormed(), description: description }
 }
+
+const swaggerResponse = (data:TSchema|null, defaultMessage?:string) => {
+    if(data) {
+        return t.Object({ data: data, message: t.String({ default: defaultMessage ?? null }) })
+    } else {
+        return t.Object({ message: t.String({ default: defaultMessage ?? null }) })
+    }
+}
+